pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
